requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0

# Scheduling
apscheduler==3.10.4
pytz==2023.3.post1

# Configuration
python-dotenv==1.0.0

# Web Scraping
beautifulsoup4==4.12.2
crawl4ai==0.9.0
lxml==4.9.3
aiohttp==3.9.1

# Data Processing
arrow==1.3.0
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
structlog==23.2.0